Version ready for update of CGM, plus several improvements
- Added CommonParams to store common parameters to the interventions in
htaReportCGM
- Updated date of BasicConfigParams to 2019
- Improved StandardSevereHypoglycemiaEvent to allow further
customization
-Now DiabetesMain prints in two different files (one for general
results, one for listeners)
- Fixed (not yet?) results of EpidemiologicView for PSA
- Fixed Unaware and Uncontrolled scenarios to share some parameters
among interventions, thus improving correlation
- Added variability to death parameters in SimpleCHDSubmodel
- Now variability in the reduction of HbA1C in Uncontrolled population
is applied as second-order uncertainty, and not first-order.
- Baseline HbA1c of Uncontrolled population constrained to min and max
values (before, I was using a Normal distribution)

+++ PSIGHOS-RETAL/src/es/ull/iis/simulation/hta/diabetes/inforeceiver/EpidemiologicView.java
@@ -146,20 +146,20 @@ public class EpidemiologicView implements ExperimentListener {
 			str.append(byAge ? (length * year) + minAge : (length *year));
 			for (int i = 0; i < interventions.size(); i++) {
 				if (byAge) {
-					str.append("\t").append(String.format(Locale.US, format, nAlivePatients[i][year] / nExperiments));					
+					str.append("\t").append(String.format(Locale.US, format, nAlivePatients[i][year] / (double)nExperiments));					
 				}
-				str.append("\t").append(String.format(Locale.US, format, nDeaths[i][year] / nExperiments));
+				str.append("\t").append(String.format(Locale.US, format, nDeaths[i][year] / (double)nExperiments));
 				for (final Named cause : nDeathsByCause.keySet()) {
-					str.append("\t").append(String.format(Locale.US, format, nDeathsByCause.get(cause)[i][year] / nExperiments));
+					str.append("\t").append(String.format(Locale.US, format, nDeathsByCause.get(cause)[i][year] / (double)nExperiments));
 				}
 				for (int j = 0; j < nMainChronic[i].length; j++) {
-					str.append("\t").append(String.format(Locale.US, format, nMainChronic[i][j][year] / nExperiments));
+					str.append("\t").append(String.format(Locale.US, format, nMainChronic[i][j][year] / (double)nExperiments));
 				}
 				for (int j = 0; j < nChronic[i].length; j++) {
-					str.append("\t").append(String.format(Locale.US, format, nChronic[i][j][year] / nExperiments));
+					str.append("\t").append(String.format(Locale.US, format, nChronic[i][j][year] / (double)nExperiments));
 				}
 				for (int j = 0; j < nAcute[i].length; j++) {
-					str.append("\t").append(String.format(Locale.US, format, nAcute[i][j][year] / nExperiments));
+					str.append("\t").append(String.format(Locale.US, format, nAcute[i][j][year] / (double)nExperiments));
 				}	
 			}
 			str.append(System.lineSeparator());
